requirements.txt
Django==4.2.2
djangorestframework==3.14.0
django-allauth==0.54.0
psycopg==3.1.9
uwsgi==2.0.21
django-redis==5.3.0
celery==5.3.1
drf-yasg==1.21.6
Pillow==9.5.0
django-imagekit==4.1.0
markdown==3.4.3
django-filter==23.2
filetype==1.2.0
django-mptt==0.14.0
django-crispy-forms==1.13.0
requests==2.31.0
django-celery-email==3.0.0
m3u8==3.5.0
django-ckeditor==6.6.1
django-debug-toolbar==4.1.0
django-login-required-middleware==0.9.0
djangorestframework-simplejwt[crypto]==5.3.1
